Create addLinks.gs
addLinks is meant to add the Google Form URLs to the "Unscrubbed List" within the Master List. These URLs link to google forms with the prefilled data for each company from the "Unscrubbed List" sheet. These links are used by the intern to scrub each company. However, they also function as unique ID codes for each company. Each URL is unique to the company, even for companies that are duplicated in the spreadsheet. To generate these URLs, the program takes the data corresponding to each company and prefills it into the google form. Then, the program actually submits the form, and gathers the resulting edit submission link, which it adds to the "Unscrubbed List" sheet. Every edit submission link is unique and leads to forms that have the submitted data prefilled, which is what's necessary for this application. Also, when the intern submits the form corresponding to the edit url, the data that the intern enters is kept within the form. Thus, if the intern wants to modify their submission, they simply re-click the link and their updated information is there. This would not hold true for prefilled links.